requirements.txt
fastapi==0.74.0
Pillow==9.0.1
uvicorn==0.12.3
aiofiles==0.6.0
python-multipart==0.0.5
jinja2==2.11.2
